refactor: getter 제거
